restructure data readers
