+++ ascat/data_readers/read_eps.py
@@ -1,12 +1,45 @@
+# Copyright (c) 2018, TU Wien, Department of Geodesy and Geoinformation
+# All rights reserved.
+
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+#    * Redistributions of source code must retain the above copyright notice,
+#      this list of conditions and the following disclaimer.
+#    * Redistributions in binary form must reproduce the above copyright
+#      notice, this list of conditions and the following disclaimer in the
+#      documentation and/or other materials provided with the distribution.
+#    * Neither the name of TU Wien, Department of Geodesy and Geoinformation
+#      nor the names of its contributors may be used to endorse or promote
+#      products derived from this software without specific prior written
+#      permission.
+
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+# ARE DISCLAIMED. IN NO EVENT SHALL TU WIEN DEPARTMENT OF GEODESY AND
+# GEOINFORMATION BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+# EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+# PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+# OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+# WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
+# OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
+# ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+
+"""
+Readers for lvl1b and lvl2 data in eps format.
+"""
+
 import os
 import fnmatch
 import lxml.etree as etree
 from tempfile import NamedTemporaryFile
 from gzip import GzipFile
 from collections import OrderedDict
+from pygeobase.io_base import ImageBase
+from pygeobase.object_base import Image
 
 import numpy as np
-import numpy.lib.recfunctions as rfn
 import datetime as dt
 import matplotlib.dates as mpl_dates
 import ascat.data_readers.templates as templates
@@ -22,6 +55,53 @@ ulong_nan = 2 ** 32 - 1
 int_nan = -2 ** 15
 uint_nan = 2 ** 16 - 1
 byte_nan = -2 ** 7
+
+
+class AscatL1bEPSImage(ImageBase):
+    def __init__(self, *args, **kwargs):
+        """
+        Initialization of i/o object.
+        """
+        super(AscatL1bEPSImage, self).__init__(*args, **kwargs)
+
+    def read(self, timestamp=None, file_format=None, **kwargs):
+        longitude, latitude, data, metadata = read_eps_l1b(
+            self.filename)
+
+        return Image(longitude, latitude, data, metadata,
+                     timestamp, timekey='jd')
+
+    def write(self, *args, **kwargs):
+        pass
+
+    def flush(self):
+        pass
+
+    def close(self):
+        pass
+
+class AscatL2EPSImage(ImageBase):
+    def __init__(self, *args, **kwargs):
+        """
+        Initialization of i/o object.
+        """
+        super(AscatL2EPSImage, self).__init__(*args, **kwargs)
+
+    def read(self, timestamp=None, file_format=None, **kwargs):
+        longitude, latitude, data, metadata = read_eps_l2(
+            self.filename)
+
+        return Image(longitude, latitude, data, metadata,
+                     timestamp, timekey='jd')
+
+    def write(self, *args, **kwargs):
+        pass
+
+    def flush(self):
+        pass
+
+    def close(self):
+        pass
 
 
 class EPSProduct(object):
