+++ ascat/data_readers/templates.py
@@ -136,3 +136,37 @@ def template_SMR__001():
     dataset = np.zeros(1, dtype=struct)
 
     return dataset
+
+
+def template_ASCRS009():
+    """
+    Generic lvl1b template. (from generic IO)
+    """
+    metadata = {'temp_name': 'ASCRS009'}
+
+    struct = np.dtype([('jd', np.double),
+                       ('sat_id', np.byte),
+                       ('abs_orbit_nr', np.uint32),
+                       ('node_num', np.uint8),
+                       ('line_num', np.uint16),
+                       ('dir', np.dtype('S1')),
+                       ('swath', np.byte),
+                       ('azif', np.float32),
+                       ('azim', np.float32),
+                       ('azia', np.float32),
+                       ('incf', np.float32),
+                       ('incm', np.float32),
+                       ('inca', np.float32),
+                       ('sigf', np.float32),
+                       ('sigm', np.float32),
+                       ('siga', np.float32),
+                       ('kpf', np.float32),
+                       ('kpm', np.float32),
+                       ('kpa', np.float32),
+                       ('num_obs', np.ubyte),
+                       ('usable_flag', np.uint8)], metadata=metadata)
+
+    dataset = np.zeros(1, dtype=struct)
+
+    return dataset
+
